cli-highlight 2.1.11 (#560)
While this version was already in the supported range, this change sets it as the minimum in order to address https://github.com/highlightjs/highlight.js/issues/2877

The older versions of cli-highlight (including 2.1.4) were still using highlight.js v9, causing consumers to receive EOL warnings when using this package. While consumers could bump the version with lockfile upgrades, this makes it explicit that we're moving away from the EOL'd versions.